requirements.txt
BeautifulSoup==3.2.1
mechanize==0.2.5
python-memcached==1.57
